test: rpc edge case test for early stream exit (#3298)
Description
---
Adds a unit test where the client requests a stream, and then drops it
while the server is streaming, then immediately makes a new request.

Motivation and Context
---
Some concerns raised over this case causing RPC to break/freeze. This tests proves that this is not the case 
but that the second request must wait for the previous stream to complete. Although this is an invalid way to
use the RPC client, this test checks the internal handling of this case.  

How Has This Been Tested?
---
Test passes