revert: fix ban header syn peer if no headers provided (#3297) (#3306)
Description
---
This reverts commit 570e2223b9443fd681f1c8395405e8aae8180d94.

Motivation and Context
---
https://github.com/tari-project/tari/pull/3297#issuecomment-913602494

Having thought about it a bit - header sync is designed to be self contained. It does not have the information required to ban a peer that "lied" in listening mode. 

Suggest that we pass through the chain metadata and peer that 'kicked' the base node into sync into header sync so 
that the peer can be banned if it fails to follow through. We should also not prevent mining when in header sync until a
split and better chain have been comfirmed by header sync.

How Has This Been Tested?
---
Existing tests pass