chore: mark broken tests and add interval print for long grpc call (#3303)
Description
---
Mark broken tests as @broken. 
Add a console print (once every 10 minutes). This will prevent the CI to kill the test.
I've added Before step that print the name of the scenario. Because now when it fails in the middle and there is no cucumber output generated. It's hard to know which test it was. Because there is only console.log available.