fix: block sync validation (#3236)
Description
---
This fixes the block sync block body validation to validate the correct items
This consolidates all the syncing helper functions
This adds in unit tests for the validators

Motivation and Context
---
Block sync needs to validate all the required steps. This also consolidates all the possible validation functions to remove duplicated pieces of code between the validators so we use a single piece to test a specific consensus item. 

How Has This Been Tested?
---
Synced a node up to the tip height 22014. 
Added new unit tests. 
Ran all integration and unit tests.