chore: bundle sqlite by default for tests (#3382)
Description
---
Yet another attempt at bundling openssl and sqlite for easy debugging/testing on Windows

Motivation and Context
---
The current feature setup means that you cannot run `cargo test` in the `wallet_ffi` or any specific folder other than the root. I tried to follow the same way that other crates do this, but allowing a user to optionally specify whether a dependency is bundled or not, and not bundle it by default, but this makes running individual tests difficult. 

Since tests are run regularly, I made the default on wallet_ffi include the bundles, but then added `--no-default-features` to the mobile builds 

How Has This Been Tested?
---
Cargo test