test: fix brokend test and remove @broken (#3385)
Description
---
I've monitored the test and it was not stuck, it was just unlucky that one of the transaction was going to be processed last. Because they are not processed in the same order as we are checking them. I've doubled the timeout value.

How Has This Been Tested?
---
Manually.