fix: update message on finding a valid share in tari_mining_node (#3408)
Description
---
When pool mining, the miner finds shares (valid solutions to a block that meets the minimum difficulty of the pool and may or may not meet the minimum difficulty to be included on the chain). The miner is also only rewarded for these shares by the pool based on the rewards system the pool uses.

Motivation and Context
---
Prevents confusion between finding a block when solo mining and finding a share when pool mining. 

How Has This Been Tested?
---
N/A