fix: Handle recovering a duplicate output
If a batch of wallet recovery queries was interrupted (either due to a connection issue or the app restarting) then it was possible for a UTXO to be rescanned. The recoverer tried to import a duplicate output which produced an error. This PR catches that case, prints a log and continues.