fix: handle recovering a duplicate output (#3426)
Description
---
If a batch of wallet recovery queries was interrupted (either due to a connection issue or the app restarting) then it was possible for a UTXO to be rescanned. The recoverer tried to import a duplicate output which produced an error. This PR catches that case, prints a log and continues.

How Has This Been Tested?
---
Manually