fix: fix flakey rust tests (#3435)
Description
---
This PR aims to fix the following flakey tests:
1. `test_txo_validation`
2. `test_transaction_resending`

The fix for 1 was to put in an explicit wait for the validation protocol to complete, on the slower CI system is seems that the validation protocol was slow to update the database state before the main test queried the balance.

The fix for 2 is to increase the cool down and resend times to make them more resilient to the slower CI environment. The current times were quite tight which was fine on a local dev machine. This might still prove to be flakey in the future.

How Has This Been Tested?
---
These are fixes to tests