fix: remove consensus::encode::Encodable impl to upgrade monero-rs 0.15.0
Monero-rs has sealed the Encodable trait which prevents it from being implemented outside of the crate itself.

This PR changes FixedByteArray to FixedByteArray<N> and removes the Encodable trait where implemented in our codebase.

It also implements serialize and deserialize for [T,N] so that [u8;MAX_ARR_SIZE] can be used for the elem field in FixedByteArray struct.