chore: improve cucumber tags and run time speed (#3439)
Description
---

This PR improves the run time of certain tests to remove `waits` and make sure they have async functions to retest waitable events.
It adds a `@base-node` tag to all base node tests.
It adds a `@wallet` tag to all wallet tests.
It adds `@`  tags to all feature files.
Its adds and remove `@critical` tags to all required tests, adding where critical functionality is not tested and removing where it is double tested and or not critical to core functionality.

Motivation and Context
---
Made sure we can run all tests as short as possible. Tests are grouped with the following tags:
`@base-node` which can be used to run tests if verification of code base node functionality is desired
`@wallet` which can be used to run tests if verification of code wallet functionality is desired
`@critical` if a short test of all core functionality is desired.
`@long-running` if really detailed long-running tests are desired. Can take hours. 

Tests without a `@critical` and `@long-running` are tests that are required for passing CI, but will take longer to run.
Running test with `--tags @base-node AND @critical` or `--tags @wallet AND @critical` completes it run time below 10 minutes.