chore: replace proxy mining with sha mining in tests (#3455)
Description
---
Replaced merge mining proxy mining in most tests with sha3 miner. 
Added base_node state to grpc for use in the tests

Motivation and Context
---
The monero merge mining proxy makes a network call to the outside which can be unreliable. The sha3 miner only uses local information which is faster as well. 

We use sleeps that in the cucumber code to ensure that base_nodes have time to sync, we can just if they are in listening state to check if we can go to the next step

How Has This Been Tested?
---
Manual cucumber tests