fix: check for previously cancelled completed ten before accepting a repeat
When handling a potentially repeated inbound transaction message there was a check for previously cancelled PENDING transactions but this did not check for a previously cancelled completed transaction.
This can occur if a transaction was completed and got rejected by the mempool. If this transaction exists then we should not add the pending transaction to our list.

This PR updates the check for previously cancelled transactions to include completed transactions.