fix: prevent race condition between block propagation and sync (#3536)
Description
---

- prevents base node from switching to sync mode before the block has a chance to arrive via propagation

Motivation and Context
---
Base node may switch to sync mode many times as the block arrives at peers, syncing a single block and returning to
listening state. It is more efficient to stay in listening state for a while to give the block chance to arrive via propagation.

How Has This Been Tested?
---
Manually tested:
- more than 1 behind tip - sync as normal
- startup 1 behind missed propagation, enters sync mode
- 1 behind, prevented from entering sync mode, block arrives shortly after 

![image](https://user-images.githubusercontent.com/1057902/140295502-8fbe59b5-257d-463c-a7fa-0c70f3a73841.png)
Prevented 3 times in a few minutes on my base node.