fix: add check for old db encryption and provide warning (#3549)
Description
---
The new CipherSeed encryption scheme changed the way in which a wallet’s encryption keys are derived from the passphrase. This is a breaking change for which a migration was non-trivial. However, if this version is run on an old database it breaks the database structure so reverting to an old version was not possible.

This PR adds a check for the old DB encryption data and before the db migration is performed produces an error with a description of the issue.

How Has This Been Tested?
---
Manually