fix: stop leak of value of recovered output (#3558)
Description
---
Updated the recovery of an output to not reuse the blinding factor to use as thescript_key, but rather generate a new key. 

Motivation and Context
---
If the blinding factor is reused as the script_key, on spending you will reveal k.G, this makes guessing v.H trivial and this leaks the k.G value. 

How Has This Been Tested?
---