chore: add node id/public key to log mdc (#3559)
Description
---
Add node id and public key to the logs, using mdc.
Since you have to propagate the mdc to every thread, I'm not 100% sure it's everywhere. But it can be easily extended if we found it missing somewhere. Also there is single point of creation of the mdc data. But I propagate everything, so if we want something else in the mdc, it will be one line change.

How Has This Been Tested?
---
Manually.