test: reduce cucumber ci to critical only (#3566)
Hopefully by reducing this CI step to critical only, we'll have more green ticks for the check and will have less to maintain regularly. 

Ideally we would want to run all of the tests, but because some of the non-critical ones fail in a flaky manner, devs have started ignoring this step and making it required takes too long.