feat: get fee for transactions for stratum transcoder (#3571)
Description
---
Adds the ability for a fee to be requested for a transaction from the console wallet via the tari stratum transcoder. It makes use of the existing grpc call get_transaction_info and essentially returns a (TxId, Fee) collection of json objects.

Motivation and Context
---
Primarily used in the Tari implementation for Miningcore.

How Has This Been Tested?
---
Manually.