feat: add GRPC call to search for utxo via commitment hex (#3666)
Added a GRPC call to search for the block a UTXO was mined in via the commitment. UI command already exists, this adds a GPRC mapping to the same command.

This fixes a UI bug that when a commitment is not found, says the the node is pruned and that the utxo was found.