feat: update faux transaction status when imported output is validated
Currently when a faucet UTXO is imported the Faux transaction that is created to track the event is created with the Imported status and reported to the FFI frontend as a Confirmed transaction. However, the output is not spendable until it has been validated. This PR changes the process so that initially the Faux transaction is reported as a Pending Inbound transaction to a FFI client and when a TXO validation is completed the transaction service will check if the output for the faux tx has become spendable at which point the Faux transaction will be marked as MinedConfirmed.

Also while writing the test for this code I thought of an easy way to make the Transaction Service testing harness much neater by adding a struct to collect all the testing interface components.