fix!: fix tx and txo validation callback bug in wallet FFI
A recent change to the transaction service events meant that the Transaction Validation callback was only being called when state changed and not on success like it used to.
It was also noted that the callbacks no longer reported when a validation failed.

This PR updates the Tx and TXO validation callbacks to be called on completion of the respective validation protocols and also report if it was a success or not. Due to the simplification of the validation protocols the callbacks were simplified so that the feedback is a boolean denoting success or failure.